requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-generativeai
duckduckgo-search
python-dotenv
